add support for multiple illegal characters in heading
The `sanitizeHeading()` function was using `replace()` with a list of strings, but that only replaces the first instance of each illegal character.  This update changes the stockIllegalSymbols to a regExp object with a global flag so that all illegal characters are removed.  I don't know how to do the same thing with user defined symbols, so I left that one as it is.